+++ torch_geometric/datasets/__init__.py
@@ -82,8 +82,8 @@ from .git_mol_dataset import GitMolDataset
 from .molecule_gpt_dataset import MoleculeGPTDataset
 from .instruct_mol_dataset import InstructMolDataset
 from .tag_dataset import TAGDataset
-from .teeth3ds import Teeth3DS
 from .city import CityNetwork
+from .teeth3ds import Teeth3DS
 
 from .dbp15k import DBP15K
 from .aminer import AMiner
@@ -203,8 +203,8 @@ homo_datasets = [
     'MoleculeGPTDataset',
     'InstructMolDataset',
     'TAGDataset',
-    'Teeth3DS',
     'CityNetwork',
+    'Teeth3DS',
 ]
 
 hetero_datasets = [