+++ torch_geometric/datasets/teeth3ds.py
@@ -17,25 +17,26 @@ from torch_geometric.data import (
 
 
 class Teeth3DS(InMemoryDataset):
-    r"""Teeth3DS+, `"An Extended Benchmark for Intra-oral 3D Scans Analysis"
-    <https://crns-smartvision.github.io/teeth3ds/>`_ paper is the first
-    comprehensive public benchmark designed to advance the field of intra-oral
-    3D scan analysis developed as part of the 3DTeethSeg 2022 and 3DTeethLand
-    2024 MICCAI challenges, aiming to drive research in teeth identification,
-    segmentation, labeling, 3D modeling, and dental landmark identification.
+    r"""The Teeth3DS+ dataset from the `"An Extended Benchmark for Intra-oral
+    3D Scans Analysis" <https://crns-smartvision.github.io/teeth3ds/>`_ paper.
+
+    This dataset is the first comprehensive public benchmark designed to
+    advance the field of intra-oral 3D scan analysis developed as part of the
+    3DTeethSeg 2022 and 3DTeethLand 2024 MICCAI challenges, aiming to drive
+    research in teeth identification, segmentation, labeling, 3D modeling,
+    and dental landmark identification.
     The dataset includes at least 1,800 intra-oral scans (containing 23,999
     annotated teeth) collected from 900 patients, covering both upper and lower
     jaws separately.
 
     Args:
-        root (str): Root directory where the dataset is stored
-                    or will be downloaded.
-        split (str): Dataset split name, e.g., "Teeth3DS".
-                     or ["3DTeethSeg22_challenge", "3DTeethLand_challenge"].
-        is_train (bool, optional): If True, use the training set;
-                                   otherwise, use the test set.
-        n_sample (int, optional): Number of points to sample from each mesh.
-                                  Default is 30000.
+        root (str): Root directory where the dataset should be saved.
+        split (str): The split name (one of :obj:`"Teeth3DS"`,
+            :obj:`"3DTeethSeg22_challenge"` or :obj:`"3DTeethLand_challenge"`).
+        train (bool, optional): If :obj:`True`, loads the training dataset,
+            otherwise the test dataset. (default: :obj:`True`)
+        num_samples (int, optional): Number of points to sample from each mesh.
+            (default: :obj:`30000`)
         transform (callable, optional): A function/transform that takes in an
             :obj:`torch_geometric.data.Data` object and returns a transformed
             version. The data object will be transformed before every access.
@@ -44,64 +45,60 @@ class Teeth3DS(InMemoryDataset):
             an :obj:`torch_geometric.data.Data` object and returns a
             transformed version. The data object will be transformed before
             being saved to disk. (default: :obj:`None`)
-        pre_filter (callable, optional): A function that takes in an
-            :obj:`torch_geometric.data.Data` object and returns a boolean
-            value, indicating whether the data object should be included in the
-            final dataset. (default: :obj:`None`)
         force_reload (bool, optional): Whether to re-process the dataset.
             (default: :obj:`False`)
     """
-
     urls = {
-        "data_part_1.zip":
+        'data_part_1.zip':
         'https://osf.io/download/qhprs/',
-        "data_part_2.zip":
+        'data_part_2.zip':
         'https://osf.io/download/4pwnr/',
-        "data_part_3.zip":
+        'data_part_3.zip':
         'https://osf.io/download/frwdp/',
-        "data_part_4.zip":
+        'data_part_4.zip':
         'https://osf.io/download/2arn4/',
-        "data_part_5.zip":
+        'data_part_5.zip':
         'https://osf.io/download/xrz5f/',
-        "data_part_6.zip":
+        'data_part_6.zip':
         'https://osf.io/download/23hgq/',
-        "data_part_7.zip":
+        'data_part_7.zip':
         'https://osf.io/download/u83ad/',
-        "train_test_split":
-        "https://files.de-1.osf.io/v1/"
-        "resources/xctdy/providers/osfstorage/?zip="
+        'train_test_split':
+        'https://files.de-1.osf.io/v1/'
+        'resources/xctdy/providers/osfstorage/?zip='
     }
 
     sample_url = {
-        "teeth3ds_sample": "https://osf.io/download/vr38s/",
+        'teeth3ds_sample': 'https://osf.io/download/vr38s/',
     }
 
     landmarks_urls = {
-        "3DTeethLand_landmarks_train.zip": 'https://osf.io/download/k5hbj/',
-        "3DTeethLand_landmarks_test.zip": 'https://osf.io/download/sqw5e/',
+        '3DTeethLand_landmarks_train.zip': 'https://osf.io/download/k5hbj/',
+        '3DTeethLand_landmarks_test.zip': 'https://osf.io/download/sqw5e/',
     }
 
     def __init__(
         self,
         root: str,
-        split: str = "Teeth3DS",
-        # [3DTeethSeg22_challenge, 3DTeethLand_challenge]
-        is_train: bool = True,
-        n_sample: int = 30000,
+        split:
+        str = 'Teeth3DS',  # [3DTeethSeg22_challenge, 3DTeethLand_challenge]
+        train: bool = True,
+        num_samples: int = 30000,
         transform: Optional[Callable] = None,
         pre_transform: Optional[Callable] = None,
-        pre_filter: Optional[Callable] = None,
         force_reload: bool = False,
     ) -> None:
-        self.mode = "training" if is_train else "testing"
+
+        self.mode = 'training' if train else 'testing'
         self.split = split
-        self.n_sample = n_sample
-        super().__init__(root, transform, pre_transform, pre_filter,
+        self.num_samples = num_samples
+
+        super().__init__(root, transform, pre_transform,
                          force_reload=force_reload)
 
     @property
     def processed_dir(self) -> str:
-        return os.path.join(self.root, f"processed_{self.split}_{self.mode}")
+        return os.path.join(self.root, f'processed_{self.split}_{self.mode}')
 
     @property
     def raw_file_names(self) -> List[str]:
@@ -109,24 +106,27 @@ class Teeth3DS(InMemoryDataset):
 
     @property
     def processed_file_names(self) -> List[str]:
-        # Directory containing train/test split files
-        split_subdir = "teeth3ds_sample" if self.split == "sample" else ""
-        split_dir = osp.join(self.raw_dir, split_subdir,
-                             f"{self.split}_train_test_split")
+        # Directory containing train/test split files:
+        split_subdir = 'teeth3ds_sample' if self.split == 'sample' else ''
+        split_dir = osp.join(
+            self.raw_dir,
+            split_subdir,
+            f'{self.split}_train_test_split',
+        )
 
         split_files = glob(osp.join(split_dir, f'{self.mode}*.txt'))
 
-        # Collect all file names from the split files
+        # Collect all file names from the split files:
         combined_list = []
         for file_path in split_files:
             with open(file_path) as file:
                 combined_list.extend(file.read().splitlines())
 
-        # Generate the list of processed file paths
-        return [f"{file_name}.pt" for file_name in combined_list]
+        # Generate the list of processed file paths:
+        return [f'{file_name}.pt' for file_name in combined_list]
 
     def download(self) -> None:
-        if self.split == "sample":
+        if self.split == 'sample':
             for key, url in self.sample_url.items():
                 path = download_url(url, self.root, filename=key)
                 extract_zip(path, self.raw_dir)
@@ -136,7 +136,6 @@ class Teeth3DS(InMemoryDataset):
                 path = download_url(url, self.root, filename=key)
                 extract_zip(path, self.raw_dir)
                 os.unlink(path)
-            # download landmarks
             for key, url in self.landmarks_urls.items():
                 path = download_url(url, self.root, filename=key)
                 extract_zip(path, self.raw_dir)  # Extract each downloaded part
@@ -145,46 +144,44 @@ class Teeth3DS(InMemoryDataset):
     def process_file(self, file_path: str) -> Optional[Data]:
         """Processes the input file path to load mesh data, annotations,
         and prepare the input features for a graph-based deep learning model.
-
-        Args:
-            file_path (str): Path to the `.obj` file representing the mesh.
-
-        Returns:
-            Data: A PyTorch Geometric Data object containing processed data.
         """
         import trimesh
         from fpsample import bucket_fps_kdline_sampling
 
-        # Load mesh
         mesh = trimesh.load_mesh(file_path)
+
         if isinstance(mesh, list):
             # Handle the case where a list of Geometry objects is returned
-            mesh = mesh[0]  # Choose the first mesh
-        # Perform sampling on mesh vertices
-        if hasattr(mesh, "vertices"):
-            vertices = mesh.vertices
-        else:
-            raise AttributeError("Mesh object has no attribute 'vertices'")
-        if hasattr(mesh, "vertex_normals"):
-            vertex_normals = mesh.vertex_normals
-        else:
-            raise AttributeError(
-                "Mesh object has no attribute 'vertex_normals'")
+            mesh = mesh[0]
 
-        if len(vertices) < self.n_sample:
-            sampled_indices = np.random.choice(len(vertices), self.n_sample,
-                                               replace=True)
+        vertices = mesh.vertices
+        vertex_normals = mesh.vertex_normals
+
+        # Perform sampling on mesh vertices:
+        if len(vertices) < self.num_samples:
+            sampled_indices = np.random.choice(
+                len(vertices),
+                self.num_samples,
+                replace=True,
+            )
         else:
             sampled_indices = bucket_fps_kdline_sampling(
-                vertices, self.n_sample, h=5, start_idx=0)
-        assert len(sampled_indices) == self.n_sample, (
-            f"Sampled points mismatch: Expected {self.n_sample},"
-            f" but got {len(sampled_indices)} for {file_path}")
-        # Extract features and annotations for the sampled points
+                vertices,
+                self.num_samples,
+                h=5,
+                start_idx=0,
+            )
+
+        if len(sampled_indices) != self.num_samples:
+            raise RuntimeError(f"Sampled points mismatch, expected "
+                               f"{self.num_samples} points, but got "
+                               f"{len(sampled_indices)} for '{file_path}'")
+
+        # Extract features and annotations for the sampled points:
         pos = torch.tensor(vertices[sampled_indices], dtype=torch.float)
         x = torch.tensor(vertex_normals[sampled_indices], dtype=torch.float)
 
-        # Load segmentation annotations
+        # Load segmentation annotations:
         seg_annotation_path = file_path.replace('.obj', '.json')
         if osp.exists(seg_annotation_path):
             with open(seg_annotation_path) as f:
@@ -198,9 +195,11 @@ class Teeth3DS(InMemoryDataset):
         else:
             y = torch.empty(0, 3)
             instances = torch.empty(0, 3)
-        # Load landmarks annotations
+
+        # Load landmarks annotations:
         landmarks_annotation_path = file_path.replace('.obj', '__kpt.json')
-        # Parse keypoint annotations into structured tensors
+
+        # Parse keypoint annotations into structured tensors:
         keypoints_dict: Dict[str, List] = {
             key: []
             for key in [
@@ -220,26 +219,28 @@ class Teeth3DS(InMemoryDataset):
                 landmarks_annotations = json.load(f)
 
             for keypoint in landmarks_annotations['objects']:
-                keypoints_dict[keypoint["class"]].extend(keypoint["coord"])
+                keypoints_dict[keypoint['class']].extend(keypoint['coord'])
 
             keypoint_tensors = {
                 k: torch.tensor(np.asarray(v),
                                 dtype=torch.float).reshape(-1, 3)
                 for k, v in keypoints_dict.items()
             }
-        # Create the PyTorch Geometric Data object
-        data = Data(pos=pos, x=x, y=y, instances=instances,
-                    jaw=file_path.split('.obj')[0].split('_')[1],
-                    mesial=keypoint_tensors['Mesial'],
-                    distal=keypoint_tensors['Distal'],
-                    cusp=keypoint_tensors['Cusp'],
-                    inner_point=keypoint_tensors['InnerPoint'],
-                    outer_point=keypoint_tensors['OuterPoint'],
-                    facial_point=keypoint_tensors['FacialPoint'])
 
-        # Apply optional pre-filter and pre-transform if specified
-        if self.pre_filter is not None and not self.pre_filter(data):
-            return None
+        data = Data(
+            pos=pos,
+            x=x,
+            y=y,
+            instances=instances,
+            jaw=file_path.split('.obj')[0].split('_')[1],
+            mesial=keypoint_tensors['Mesial'],
+            distal=keypoint_tensors['Distal'],
+            cusp=keypoint_tensors['Cusp'],
+            inner_point=keypoint_tensors['InnerPoint'],
+            outer_point=keypoint_tensors['OuterPoint'],
+            facial_point=keypoint_tensors['FacialPoint'],
+        )
+
         if self.pre_transform is not None:
             data = self.pre_transform(data)
 
@@ -247,21 +248,21 @@ class Teeth3DS(InMemoryDataset):
 
     def process(self) -> None:
         for file in tqdm(self.processed_file_names):
-            file_name = file.split('.')[0]
-            file_path = glob(
-                osp.join(self.raw_dir, '**', '*', '*', file_name + '.obj'))
-            if len(file_path) == 1:
-                data = self.process_file(file_path[0])
+            name = file.split('.')[0]
+            path = osp.join(self.raw_dir, '**', '*', name + '.obj')
+            paths = glob(path)
+            if len(paths) == 1:
+                data = self.process_file(paths[0])
                 torch.save(data, osp.join(self.processed_dir, file))
 
     def len(self) -> int:
         return len(self.processed_file_names)
 
     def get(self, idx: int) -> Data:
-        file = self.processed_file_names[idx]
-        data = torch.load(osp.join(self.processed_dir, file),
-                          weights_only=False)
-        return data
+        return torch.load(
+            osp.join(self.processed_dir, self.processed_file_names[idx]),
+            weights_only=False,
+        )
 
     def __repr__(self) -> str:
         return (f'{self.__class__.__name__}({len(self)}, '